Now working :D

+++ src/main/java/org/cubixmc/server/network/packets/play/PacketInPlayerPosition.java
@@ -17,9 +17,9 @@ public class PacketInPlayerPosition extends PacketIn {
 
     @Override
     public void decode(Codec codec) {
-        this.x = codec.readFloat();
-        this.feetY = codec.readFloat();
-        this.z = codec.readFloat();
+        this.x = codec.readDouble();
+        this.feetY = codec.readDouble();
+        this.z = codec.readDouble();
         this.onGround = codec.readBool();
     }
